Cors sredjen

+++ src/main/java/rs/raf/filipdraganicrn5417/avioraf/model/AvionskaKompanija.java
@@ -10,7 +10,7 @@ public class AvionskaKompanija {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id; //mora biti jedinstven, ne prikazuje se na frontendu
+    private long id; //mora biti jedinstven, ne prikazuje se na frontendu
 
     @Column(nullable = false, unique = true)
     private String name; //mora biti jedinstveno i mora se sastojati od alfanumerickih karaktera
