+++ src/main/java/rs/raf/filipdraganicrn5417/avioraf/model/Grad.java
@@ -11,7 +11,7 @@ public class Grad {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id; //mora biti jedinstven, ne prikazuje se na frontendu
+    private long id; //mora biti jedinstven, ne prikazuje se na frontendu
 
     @Column(nullable = false, unique = true)
     private String name; //ime grada, mora biti jedinstveno
